fixed DRL agent and add data collection
