+++ proyecto_final/code/output_drl.json
@@ -1,0 +1,12 @@
+{
+  "0": {
+    "episode": 0,
+    "reward": 0,
+    "time": 0.031378746032714844
+  },
+  "1": {
+    "episode": 1,
+    "reward": 0,
+    "time": 0.05539417266845703
+  }
+}
