--- proyecto_final/code/q_learning/drl.py
@@ -1,170 +1,0 @@
-from snake_env import Snake
-
-import random
-import numpy as np
-from keras import Sequential
-from collections import deque
-from keras.layers import Dense
-import matplotlib.pyplot as plt
-from keras.optimizers import adam_v2
-from plot_script import plot_result
-import time
-import json
-
-
-class DQN:
-
-    """ Deep Q Network """
-
-    def __init__(self, env, params):
-
-        self.action_space = env.action_space
-        self.state_space = env.state_space
-        self.epsilon = params['epsilon'] 
-        self.gamma = params['gamma'] 
-        self.batch_size = params['batch_size'] 
-        self.epsilon_min = params['epsilon_min'] 
-        self.epsilon_decay = params['epsilon_decay'] 
-        self.learning_rate = params['learning_rate']
-        self.layer_sizes = params['layer_sizes']
-        self.memory = deque(maxlen=2500)
-        self.model = self.build_model()
-
-
-    def build_model(self):
-        model = Sequential()
-        for i in range(len(self.layer_sizes)):
-            if i == 0:
-                model.add(Dense(self.layer_sizes[i], input_shape=(self.state_space,), activation='relu'))
-            else:
-                model.add(Dense(self.layer_sizes[i], activation='relu'))
-        model.add(Dense(self.action_space, activation='softmax'))
-        model.compile(loss='mse', optimizer=adam_v2(lr=self.learning_rate))
-        return model
-
-
-    def remember(self, state, action, reward, next_state, done):
-        self.memory.append((state, action, reward, next_state, done))
-
-
-    def act(self, state):
-
-        if np.random.rand() <= self.epsilon:
-            return random.randrange(self.action_space)
-        act_values = self.model.predict(state)
-        return np.argmax(act_values[0])
-
-
-    def replay(self):
-
-        if len(self.memory) < self.batch_size:
-            return
-
-        minibatch = random.sample(self.memory, self.batch_size)
-        states = np.array([i[0] for i in minibatch])
-        actions = np.array([i[1] for i in minibatch])
-        rewards = np.array([i[2] for i in minibatch])
-        next_states = np.array([i[3] for i in minibatch])
-        dones = np.array([i[4] for i in minibatch])
-
-        states = np.squeeze(states)
-        next_states = np.squeeze(next_states)
-
-        targets = rewards + self.gamma*(np.amax(self.model.predict_on_batch(next_states), axis=1))*(1-dones)
-        targets_full = self.model.predict_on_batch(states)
-
-        ind = np.array([i for i in range(self.batch_size)])
-        targets_full[[ind], [actions]] = targets
-
-        self.model.fit(states, targets_full, epochs=1, verbose=0)
-        if self.epsilon > self.epsilon_min:
-            self.epsilon *= self.epsilon_decay
-
-class data_collected:
-    def __init__(self,episode,reward,time):
-        self.episode = episode
-        self.reward = reward
-        self.time = time
-    
-    def __str__(self) -> str:
-        res = 'Ep : '+str(self.episode)+' | '+str(self.reward)+' puntos en '+str(self.time)+' seg.'
-        return res
-
-
-def train_dqn(episode, env):
-
-    rewards_per_episode = list()
-    sum_of_rewards = []
-    agent = DQN(env, params)
-    for e in range(episode):
-        start = time.time()
-        score_episode = 0
-        state = env.reset()
-        state = np.reshape(state, (1, env.state_space))
-        score = 0
-        max_steps = 10000
-        for i in range(max_steps):
-            action = agent.act(state)
-            # print(action)
-            prev_state = state
-            next_state, reward, done, new_score_episode = env.step(action)
-            score += reward
-            next_state = np.reshape(next_state, (1, env.state_space))
-            agent.remember(state, action, reward, next_state, done)
-            state = next_state
-            if params['batch_size'] > 1:
-                agent.replay()
-            if done:
-                print(f'final state before dying: {str(prev_state)}')
-                print(f'episode: {e+1}/{episode}, score: {score}')
-                break
-            else:
-                score_episode = new_score_episode
-        sum_of_rewards.append(score)
-        end = time.time()
-        dc = data_collected(e,score_episode,end-start)
-        rewards_per_episode.append(dc)
-    return sum_of_rewards, rewards_per_episode
-
-
-if __name__ == '__main__':
-
-    params = dict()
-    params['name'] = None
-    params['epsilon'] = 1
-    params['gamma'] = .95
-    params['batch_size'] = 500
-    params['epsilon_min'] = .01
-    params['epsilon_decay'] = .995
-    params['learning_rate'] = 0.00025
-    params['layer_sizes'] = [128, 128, 128]
-
-    results = dict()
-    ep = 50
-
-    # for batchsz in [1, 10, 100, 1000]:
-    #     print(batchsz)
-    #     params['batch_size'] = batchsz
-    #     nm = ''
-    #     params['name'] = f'Batchsize {batchsz}'
-    env_infos = {'States: only walls':{'state_space':'no body knowledge'}, 'States: direction 0 or 1':{'state_space':''}, 'States: coordinates':{'state_space':'coordinates'}, 'States: no direction':{'state_space':'no direction'}}
-
-    # for key in env_infos.keys():
-    #     params['name'] = key
-    #     env_info = env_infos[key]
-    #     print(env_info)
-    #     env = Snake(env_info=env_info)
-    env = Snake()
-    sum_of_rewards, export = train_dqn(ep, env)
-    results[params['name']] = sum_of_rewards
-    
-    plot_result(results, direct=True, k=20)
-    
-    data_dump = dict()
-    for i in range(ep):
-        data_dump [i] = export[i].__dict__
-    out_file = open("./ia-uncuyo-2021/proyecto_final/code/output_drl.json", "w") 
-        
-    json.dump(data_dump, out_file, indent = 2) 
-        
-    out_file.close()
